Update initial page render function
At this point I've got the main heading, the list container, and the task container rendering on page load. I'm using a 6x6 grid for the page layout, although I may update that to a 12x12 for more flexibility. Right now, the tast list is too large imo when spanning 2 columns and too small when spanning 1. It may look better spanning 3 / 12.